Create RODOS.cpp
RODOS and ROS for sensors in c++ are used in robotics industry and aerospace. To configure structural robotics inside the rockets and the satellites we need to take in consideration apart from the aerodynamics systems, the nucleus in RODOS, ROS
and c++ signals, we need also dns signals and satellite signals for aerospace projects and test with a debugger or strace the
memory and the pointers of our satellite-sensors and server-signals for robotic systems and aerospace research.